# schemas/generated/orchestration-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tasker Web API",
    "description": "REST API for the Tasker workflow orchestration system",
    "contact": {
      "name": "Tasker Systems",
      "email": "support@tasker-systems.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "API version 1"
    }
  ],
  "paths": {
...
    "/v1/templates": {
      "get": {
        "tags": [
          "templates"
        ],
        "summary": "List available templates",
        "description": "GET /v1/templates\n\nReturns a list of all registered task templates, optionally filtered by namespace.\nIncludes namespace summaries and template details.\n\n**Required Permission:** `templates:read`",
        "operationId": "list_templates",
        "parameters": [
          {
            "name": "namespace",
            "in": "query",
            "description": "Filter by namespace",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of available templates",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key_auth": []
          }
        ],
        "x-required-permission": "templates:read"
      }
    },
    "/v1/templates/{namespace}/{name}/{version}": {
      "get": {
        "tags": [
          "templates"
        ],
        "summary": "Get a specific template by namespace, name, and version",
        "description": "GET /v1/templates/{namespace}/{name}/{version}\n\nReturns detailed information about a specific template including its step definitions.\n\n**Required Permission:** `templates:read`",
        "operationId": "get_template",
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "description": "Template namespace",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Template name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "in": "path",
            "description": "Template version",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Template details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateDetail"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Template not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key_auth": []
          }
        ],
        "x-required-permission": "templates:read"
      }
    }
  },
  "components": {
    "schemas": {
      "ApiError": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "NotFound"
            ],
            "properties": {
              "NotFound": {
                "type": "object",
                "required": [
                  "message"
                ],
                "properties": {
                  "message": {
                    "type": "string"
                  }
                }
              }
            }
          },
          {
            "type": "string",
            "enum": [
              "Forbidden"
            ]
          },
          {
            "type": "string",
            "enum": [
              "Unauthorized"
            ]
          },
          {
            "type": "object",
            "required": [
              "BadRequest"
            ],
            "properties": {
              "BadRequest": {
                "type": "object",
                "required": [
                  "message"
                ],
                "properties": {
                  "message": {
                    "type": "string"
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "description": "TAS-154: Conflict - resource already exists (e.g., duplicate identity hash)",
            "required": [
              "Conflict"
            ],
            "properties": {
              "Conflict": {
                "type": "object",
                "description": "TAS-154: Conflict - resource already exists (e.g., duplicate identity hash)",
                "required": [
                  "message"
                ],
                "properties": {
                  "message": {
                    "type": "string"
                  }
                }
              }
            }
          },
          {
            "type": "string",
            "enum": [
              "ServiceUnavailable"
            ]
          },
          {
            "type": "string",
            "enum": [
              "Timeout"
            ]
          },
          {
            "type": "string",
            "enum": [
              "CircuitBreakerOpen"
            ]
          },
          {
            "type": "object",
            "description": "TAS-75: System backpressure active - includes Retry-After header",
            "required": [
              "Backpressure"
            ],
            "properties": {
              "Backpressure": {
                "type": "object",
                "description": "TAS-75: System backpressure active - includes Retry-After header",
                "required": [
                  "reason",
                  "retry_after_seconds"
                ],
                "properties": {
                  "reason": {
                    "type": "string"
                  },
                  "retry_after_seconds": {
                    "type": "integer",
                    "format": "int64",
                    "minimum": 0
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": [
              "DatabaseError"
            ],
            "properties": {
              "DatabaseError": {
                "type": "object",
                "required": [
                  "operation"
                ],
                "properties": {
                  "operation": {
                    "type": "string"
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": [
              "AuthenticationError"
            ],
            "properties": {
              "AuthenticationError": {
                "type": "object",
                "required": [
                  "reason"
                ],
                "properties": {
                  "failure_context": {
                    "oneOf": [
                      {
                        "type": "null"
                      },
                      {
                        "$ref": "#/components/schemas/AuthFailureContext",
                        "description": "Machine-readable failure context for gateway signaling headers"
                      }
                    ]
                  },
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": [
              "AuthorizationError"
            ],
            "properties": {
              "AuthorizationError": {
                "type": "object",
                "required": [
                  "reason"
                ],
                "properties": {
                  "failure_context": {
                    "oneOf": [
                      {
                        "type": "null"
                      },
                      {
                        "$ref": "#/components/schemas/AuthFailureContext",
                        "description": "Machine-readable failure context for gateway signaling headers"
                      }
                    ]
                  },
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": [
              "InvalidUuid"
            ],
            "properties": {
              "InvalidUuid": {
                "type": "object",
                "required": [
                  "uuid"
                ],
                "properties": {
                  "uuid": {
                    "type": "string"
                  }
                }
              }
            }
          },
          {
            "type": "string",
            "enum": [
              "JsonError"
            ]
          },
          {
            "type": "string",
            "enum": [
              "Internal"
            ]
          }
        ],
        "description": "Web API specific errors with HTTP status code mappings"
      },
      "AuthFailureContext": {
        "type": "object",
        "description": "Auth failure context for gateway signaling.\n\nIncluded in 401/403 response headers so upstream gateways (ALB, NLB, API Gateway)\ncan make informed rate-limiting decisions based on failure severity.",
        "required": [
          "reason",
          "severity"
        ],
        "properties": {
          "reason": {
            "$ref": "#/components/schemas/AuthFailureReason",
            "description": "Machine-readable failure category"
          },
          "severity": {
            "$ref": "#/components/schemas/AuthFailureSeverity",
            "description": "Suggested severity for gateway action"
          }
        }
      },
      "AuthFailureReason": {
        "type": "string",
        "description": "Machine-readable auth failure reasons.",
        "enum": [
          "Missing",
          "Expired",
          "Invalid",
          "Malformed",
          "Forbidden"
        ]
      },
      "AuthFailureSeverity": {
        "type": "string",
        "description": "Severity levels for auth failures (signals gateway rate-limiting behavior).",
        "enum": [
          "Low",
          "Medium",
          "High"
        ]
      },
...
      "NamespaceSummary": {
        "type": "object",
        "description": "Summary of a namespace with its templates",
        "required": [
          "name",
          "template_count"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional namespace description"
          },
          "name": {
            "type": "string",
            "description": "Namespace name"
          },
          "template_count": {
            "type": "integer",
            "description": "Number of templates in this namespace",
            "minimum": 0
          }
        }
      },
...
      "StepDefinition": {
        "type": "object",
        "description": "Definition of a step within a template",
        "required": [
          "name",
          "default_retryable",
          "default_max_attempts"
        ],
        "properties": {
          "default_max_attempts": {
            "type": "integer",
            "format": "int32",
            "description": "Maximum retry attempts"
          },
          "default_retryable": {
            "type": "boolean",
            "description": "Whether the step is retryable by default"
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Step description"
          },
          "name": {
            "type": "string",
            "description": "Step name/handler identifier"
          }
        }
      },
...
      "TemplateDetail": {
        "type": "object",
        "description": "Detailed information about a single template\n\nReturned when fetching a specific template by namespace/name/version.",
        "required": [
          "name",
          "namespace",
          "version",
          "configuration",
          "steps"
        ],
        "properties": {
          "configuration": {
            "type": "object",
            "description": "Template configuration (JSON)"
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional description"
          },
          "name": {
            "type": "string",
            "description": "Template name"
          },
          "namespace": {
            "type": "string",
            "description": "Namespace the template belongs to"
          },
          "steps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StepDefinition"
            },
            "description": "Step definitions for this template"
          },
          "version": {
            "type": "string",
            "description": "Template version"
          }
        }
      },
      "TemplateListResponse": {
        "type": "object",
        "description": "Response for template listing endpoint\n\nProvides an overview of all available templates, optionally filtered by namespace.",
        "required": [
          "namespaces",
          "templates",
          "total_count"
        ],
        "properties": {
          "namespaces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NamespaceSummary"
            },
            "description": "Summary of available namespaces"
          },
          "templates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TemplateSummary"
            },
            "description": "List of templates (optionally filtered)"
          },
          "total_count": {
            "type": "integer",
            "description": "Total number of templates (before filtering)",
            "minimum": 0
          }
        }
      },
      "TemplateSummary": {
        "type": "object",
        "description": "Summary information about a template\n\nUsed in list responses to provide an overview of available templates.",
        "required": [
          "name",
          "namespace",
          "version",
          "step_count"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional description"
          },
          "name": {
            "type": "string",
            "description": "Template name"
          },
          "namespace": {
            "type": "string",
            "description": "Namespace the template belongs to"
          },
          "step_count": {
            "type": "integer",
            "description": "Number of steps in the template",
            "minimum": 0
          },
          "version": {
            "type": "string",
            "description": "Template version"
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key_auth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}